+++ otelconf/v0.3.0/config.go
@@ -222,9 +222,13 @@ func createHeadersConfig(headers []NameStringValuePair, headersList *string) (ma
 	// Headers take precedence over HeadersList, so this has to be after HeadersList is processed
 	if len(headers) > 0 {
 		for _, kv := range headers {
-			if kv.Value != nil {
-				result[kv.Name] = *kv.Value
+			if kv.Value == nil {
+				continue
 			}
+			if kv.Name == "" {
+				return nil, errors.New("invalid header: empty name")
+			}
+			result[kv.Name] = *kv.Value
 		}
 	}
 	return result, nil