+++ otelconf/v0.3.0/config_test.go
@@ -687,6 +687,16 @@ func TestCreateHeadersConfig(t *testing.T) {
 			headersList: ptr("==="),
 			wantErr:     "invalid headers list: invalid key: \"\"",
 		},
+		{
+			name: "empty header name",
+			headers: []NameStringValuePair{
+				{
+					Name:  "",
+					Value: ptr("token"),
+				},
+			},
+			wantErr: "invalid header: empty name",
+		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {