+++ otelconf/x/config_common.go
@@ -301,9 +301,13 @@ func createHeadersConfig(headers []NameStringValuePair, headersList *string) (ma
 	}
 	// Headers take precedence over HeadersList, so this has to be after HeadersList is processed.
 	for _, kv := range headers {
-		if kv.Value != nil {
-			result[kv.Name] = *kv.Value
+		if kv.Value == nil {
+			continue
 		}
+		if kv.Name == "" {
+			return nil, newErrInvalid("invalid header: empty name")
+		}
+		result[kv.Name] = *kv.Value
 	}
 	return result, nil
 }